+++ src/App.tsx
@@ -1,10 +1,9 @@
-import {URLs} from "../src/constants/urls";
+import {URLs} from "./utils/constants/constants";
 import styles from "./App.module.css";
 import { Routes, Route } from "react-router-dom";
-import { SIGNUP } from "./utils/constants/constants";
 import Signup from "./pages/signup/signup";
-import Header from "./pages/main/header/header";
-import Footer from "./pages/main/footer/footer";
+import Header from "./components/header/header";
+import Footer from "./components/footer/footer";
 import Content from "./pages/main/content/content";
 
 function App() {
