+++ src/components/header/navBar/NavBar.tsx
@@ -1,4 +1,4 @@
-import {URLs} from "../../../../constants/urls";
+import {URLs} from "../../../utils/constants/constants";
 import navbarStyle from "./NavBar.module.css";
 import { NavLink } from "react-router-dom";
