--- src/constants/urls.ts
@@ -1,4 +1,0 @@
-  export const URLs = {
-      HOME_PAGE: '/',
-      SIGN_UP: `/signup`
-  } as const;
